--- Navigation/Navigation/ProfileVC.swift
@@ -1,27 +1,0 @@
-//
-//  ProfileVC.swift
-//  Navigation
-//
-//  Created by Yan Sakhnevich on 16.10.2021.
-//
-
-import UIKit
-
-class ProfileVC: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        title = "Profile"
-        view.backgroundColor = .systemCyan
-        let appearance = UINavigationBarAppearance()
-        appearance.backgroundColor = .systemBackground
-        appearance.titleTextAttributes = [.foregroundColor: UIColor.black]
-        appearance.largeTitleTextAttributes = [.foregroundColor: UIColor.white]
-        navigationController?.navigationBar.standardAppearance = appearance
-        navigationController?.navigationBar.compactAppearance = appearance
-        navigationController?.navigationBar.scrollEdgeAppearance = appearance
-    
-    }
-
-
-}
